backend/requirements/dev.txt
-r base.txt
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
factory-boy==3.3.1
ipython==8.31.0
django-debug-toolbar==5.0.1
